% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOpair.R
\name{GOpair}
\alias{GOpair}
\title{GOpair is a function to compare two mygo objects. it compares the common genes and also returns unique genes for the object}
\usage{
GOpair(mygo1, mygo2)
}
\description{
GOpair is a function to compare two mygo objects. it compares the common genes and also returns unique genes for the object
}

