% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOsummer.R
\name{GOsummer}
\alias{GOsummer}
\title{GOsummer is the function that you can use to do summerization on the myGO object.}
\usage{
GOsummer(mygo, Type, Term)
}
\arguments{
\item{mygo, Type, Term}{Type is the ontology type: MF, BP, CC, PC
Term is the GOterms or terms you are interested to see, such as response, binding, et.

MF terms include: "transporter activity", "binding", "catalytic activity", "structural molecular activity", "receptor activity", "antioxidant activity"

BP terms include:"cellular process", "localization", "metabolic process", "response to stimulus", "biological regulation", "cellular process", "cellular component organization or biogenesis"

CC terms include: "cell part", "marcromolecular complex", "membrane", "organelle"

PC terms include:"chaperone", "hydrolase","transporter", "oxidoreductase", "ligase", "enzyme modulator", "isomerase", "nucleic acid binding", "receptor", "transcription factor", "transfer/carrier protein", "transferase", "defense/immunity protein", "ribosomal RNA",  "tRNA"

IF inputs are not in Type or Term, it returns a list of terms you can choose from

IF Type input is not in the type vector, and you know the Term, it returns a list of summaries based on the Term

IF Term is unknown, and you know the Type, it returns a list of summaries based on the Type

IF both inputs are in the function, it returns a list of summaries according to the Type and Term}
}
\description{
GOsummer is the function that you can use to do summerization on the myGO object.
}
\examples{

\dontrun{GOsummer(mygo.e.34hr, "MF", "transporter activity")}
\dontrun{GOsummer(mygo.e.34hr, "I don't know", "transporter")}
}

