% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGO.R
\name{myGO}
\alias{myGO}
\title{A myGO function}
\usage{
myGO(MF, BP, CC, PC, annoset, write)
}
\description{
This allows you to do Deinococcus radiodurans ontology analysis. The funciton generates a data frame with gene symbol, refID, gene ID, Differential expression pattern, annotation from NCBI, GO terms from Panther.
}
\details{
The data base is based on all the genes that have hits in PATHNER from our previous analysis

You need MF, BP, CC, PC for database construction for this function to run

You need to input the annotation dataset from annotation.rmd

write = T to write a csv file with all the genes in the ontology database
}
\examples{
\dontrun{myGO(MF, BP, CC, PC, edgeR_34hr, FALSE)}

}

